python-to-c-converter/requirements.txt
openai>=1.0.0
gradio==4.44.0
python-dotenv>=1.0.0